Move `Bastion` component to `pkg/component`
