+++ pkg/component/extensions/bastion/bastion_suite_test.go
@@ -1,0 +1,17 @@
+// SPDX-FileCopyrightText: SAP SE or an SAP affiliate company and Gardener contributors
+//
+// SPDX-License-Identifier: Apache-2.0
+
+package bastion_test
+
+import (
+	"testing"
+
+	. "github.com/onsi/ginkgo/v2"
+	. "github.com/onsi/gomega"
+)
+
+func TestBastion(t *testing.T) {
+	RegisterFailHandler(Fail)
+	RunSpecs(t, "Component Extensions Bastion Suite")
+}
