+++ skaffold-gardenadm.yaml
@@ -70,6 +70,7 @@ build:
             - pkg/component/etcd/etcd
             - pkg/component/etcd/etcd/constants
             - pkg/component/extensions/backupentry
+            - pkg/component/extensions/bastion
             - pkg/component/extensions/containerruntime
             - pkg/component/extensions/controlplane
             - pkg/component/extensions/crds
